refector modules to break dependencies
Gemini JPA does refresh bundles when the configuration is applied in
order to weave the classes property. this will cause the depending
bundles to be stopped and started and their depending bundles as well.

this causes the web ui to shut down while configuring the database
connection, which could cause issues in the web ui itself.

now these ties are broken up so that refreshing the JPA unit bundles
will not stop and start everything.