make the database update a bit more complex and a bit less drastic
The process not actually checks if there is a difference in the
configuration and waits for the initial update from the
configuration admin.

this reduces the number of times the configuration is updated and
therefore Gemini JPA needs to reconfigure, which reduces the number of
times bundles are refreshed and services being restarted.

During a normal startup there should be no configuration action at all.