show an artifact label and improve loading of meta data
this change introduces the concept of a common artifact label, which
can be provided independent of the namespace. The web UI will show
labels for all artifact labels of all namespaces.

In addition the loading of meta data has been improved so that all
artifacts and meta data can be loaded without causing an n+1 select
issue, by merging the data in java code. this allows using the
meta data in the tree view and improves loading speed of the OSGi
bundle view.