fix a problem aggregating the channels
Aggregating the channels missed out child artifacts due to the fact
that they got not initialized properly when the parent/child relation
was set. Read from the database the relation was intact, but in the
same session the relation was not intanct.

There were two possible options, clear the entity manager and force
a reload from the database or fix the original problem. However this
also required to remove the "id" variants of child and parent, since
the ID is not known when the relation is set.