create a maven directory structure for extracted POM files, #71
This commit adds an artifact to the maven repository structure if
there is a child-artifact element which is an extracted POM file.

It takes GAV from the POM file and the extension is the extension of
the artifact name.

This is the last commit which enables uploading JAR files, which
contain a "pom" and re-created the maven repository structure, allowing
maven read access to this artifact.