remove the requirement for Extractor to return its own aspect
In the past the Extractor interface was extending the
ChannelAspectFunction interface in order to return its own aspect.
This was done for the sole purpose of getting the aspect factory id
of an extractor. However the aspect processing now has a different way
to do this. So this requirement is now removed completely.