lock the channel around the handler operation, #97
The issue is caused by the fact that the channel cache entries
are written to the database, but are not committed. The second
operation gets woken up by the lock manager too early. The first
operation opens the transaction before locking the channel, which,
as of now, also closes the transaction _after_ releasing the lock.