rework streamCache operation to return if the entry was found
Instead of throwing a FileNotFoundException the return code now
indicates that the cache entry was not found.

This was necessary since twice the maven repository adapter broke
due to the fact that the service method threw a FileNotFoundException
wrapped in some other exception. This time due to the new profiler.