fix the issue of xalan being used first in the new setup, #110
It seems that using the new setup Xalan gets found first and then
gets cached by the internal service lookup mechanism. So Java
remembers to use Xalan, which might not be accesible in other bundles
though.

As a workaround the XmlHelper will now add a dependency to Xalan and
set the TCCL accordingly. So as long as XmlHelper is used, there
should not be an issue anymore.