build new DEB and RPM package based on "org.eclipse.packagedrone"
This change will change the DEB and RPM creation in a way that it will
create packages with the name "org.eclipse.packagedrone.server" now
instead of "package-drone-server". This will not only reflect the change
towards moving to the Eclipse Foundation, but also allows us to declare
a package conflict to "package-drone-server" which features the old
storage model. Allowing a seamless package upgrade would result in a
broken system installation otherwise.

In addition the RPM release numbers will be generated differently now so
that SNAPSHOTS have the lowest priority, the milestones (m), then
release candicates (rc) and finally the actual release builds.

Also has there been a dependency declared on "jre >= 1.8.0" for CentOS
7, since they do provide an OpenJDK 1.8.0 and JRE (also headless mode)
now.